+++ src/usePreloadedQuery.ts
@@ -1,7 +1,6 @@
 import { useState, useEffect } from 'react';
 import { OperationType } from 'relay-runtime';
-import { LoadQuery } from './loadQuery';
-import { RenderProps } from './RelayHooksType';
+import { RenderProps, LoadQuery } from './RelayHooksType';
 import { useRelayEnvironment } from './useRelayEnvironment';
 
 export const usePreloadedQuery = <TOperationType extends OperationType = OperationType>(
